Rag is exposed as a reference
