+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/testing/unit/FakePromptRunner.kt
@@ -20,9 +20,6 @@ import com.embabel.agent.core.ToolGroup
 import com.embabel.agent.core.ToolGroupRequirement
 import com.embabel.agent.core.support.safelyGetToolCallbacks
 import com.embabel.agent.prompt.element.ContextualPromptElement
-import com.embabel.agent.rag.RagService
-import com.embabel.agent.rag.tools.RagOptions
-import com.embabel.agent.rag.tools.SingleShotRagServiceSearchTools
 import com.embabel.agent.spi.InteractionId
 import com.embabel.agent.spi.LlmInteraction
 import com.embabel.chat.Message
@@ -191,12 +188,6 @@ data class FakePromptRunner(
             templateRenderer = JinjavaTemplateRenderer(),
             promptRunnerOperations = this,
         )
-    }
-
-    override fun withRag(options: RagOptions): PromptRunner {
-        logger.warn("RAG tools not implemented in FakePromptRunner")
-        return this.withToolObject(SingleShotRagServiceSearchTools(RagService.empty(), RagOptions()))
-
     }
 
     override fun withHandoffs(vararg outputTypes: Class<*>): PromptRunner {
